requirements.txt
Django
psycopg2-binary
pytz
requests
python-barcode
reportlab
svglib